refactor: rename app packages
