+++ src/main/java/no/fdk/sparqlservice/configuration/KafkaConsumerConfig.java
@@ -1,4 +1,4 @@
-package no.fdk.configuration;
+package no.fdk.sparqlservice.configuration;
 
 
 import no.fdk.concept.ConceptEvent;
